refact(ForverPage): changed file structure for ForeverPage
