+++ apps/onboarding/src/components/ForeverPage/usePrintCodeEffect.ts
